+++ packages/react/src/router/v6/FaroRoutes.tsx
@@ -1,12 +1,12 @@
-import { useEffect, useMemo } from 'react';
+import { useEffect, useMemo, useRef } from 'react';
 
-import { globalObject } from '@grafana/faro-web-sdk';
+import { EVENT_ROUTE_CHANGE, globalObject } from '@grafana/faro-web-sdk';
 
 import { api } from '../../dependencies';
 import { NavigationType } from '../types';
 
 import { createRoutesFromChildren, isInitialized, Routes, useLocation, useNavigationType } from './routerDependencies';
-import type { ReactRouterV6RoutesProps } from './types';
+import type { EventRouteTransitionAttributes, ReactRouterV6RoutesProps } from './types';
 import { getRouteFromLocation } from './utils';
 
 export function FaroRoutes(props: ReactRouterV6RoutesProps) {
@@ -15,12 +15,28 @@ export function FaroRoutes(props: ReactRouterV6RoutesProps) {
 
   const routes = useMemo(() => createRoutesFromChildren?.(props.children) ?? [], [props.children]);
 
+  const lastRouteRef = useRef<EventRouteTransitionAttributes>({});
+
   useEffect(() => {
     if (isInitialized && (navigationType === NavigationType.Push || navigationType === NavigationType.Pop)) {
-      api.pushEvent('routeChange', {
-        route: getRouteFromLocation(routes, location),
-        url: globalObject.location?.href,
+      const route = getRouteFromLocation(routes, location);
+      const url = globalObject.location?.href;
+      const { fromRoute, fromUrl } = lastRouteRef.current;
+
+      if (route === fromRoute && url === fromUrl) {
+        return;
+      }
+
+      api.pushEvent(EVENT_ROUTE_CHANGE, {
+        toRoute: route,
+        toUrl: globalObject.location?.href,
+        ...lastRouteRef.current,
       });
+
+      lastRouteRef.current = {
+        fromRoute: route,
+        fromUrl: url,
+      };
     }
   }, [location, navigationType, routes]);